requirements.txt
os
streamlit
streamlit_chat
langchain
sentence_transformers
openai
unstructured
pinecone-client
langchain_community
